+++ Areas/Admin/Controllers/ProductController.cs
@@ -115,6 +115,11 @@ namespace ShoppingFood.Areas.Admin.Controllers
 
             var existProduct = await _dataContext.Products.FindAsync(model.Id);
 
+            if (existProduct == null)
+            {
+                _notyf.Error("Product Not Found");
+            }
+
             if (ModelState.IsValid)
             {
                 if (model.ImageUpload != null)
@@ -122,17 +127,20 @@ namespace ShoppingFood.Areas.Admin.Controllers
                     string uploadsDir = Path.Combine(_webHostEnvironment.WebRootPath, "media/products");
                     string imageName = Guid.NewGuid().ToString() + "_" + model.ImageUpload.FileName;
                     string filePath = Path.Combine(uploadsDir, imageName);
-                    string oldFilePath = Path.Combine(uploadsDir, existProduct.Image);
-                    try
+                    if (!string.IsNullOrEmpty(existProduct.Image))
                     {
+                        string oldFilePath = Path.Combine(uploadsDir, existProduct.Image);
                         if (System.IO.File.Exists(oldFilePath))
                         {
-                            System.IO.File.Delete(oldFilePath);
+                            try
+                            {
+                                System.IO.File.Delete(oldFilePath);
+                            }
+                            catch (Exception ex)
+                            {
+                                ModelState.AddModelError("", "Không thể xóa ảnh cũ: " + ex.Message);
+                            }
                         }
-                    }
-                    catch (Exception ex)
-                    {
-                        ModelState.AddModelError("Error", ex.Message);
                     }
                     FileStream fileStream = new FileStream(filePath, FileMode.Create);
                     await model.ImageUpload.CopyToAsync(fileStream);
